7/24 18:00 commit

+++ src/main/java/com/sbdc/board/repository/BoardRepository_save.java
@@ -5,10 +5,11 @@ import org.hibernate.mapping.List;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.http.converter.json.GsonBuilderUtils;
 import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BoardRepository_save extends JpaRepository<TB_BOARD_SAVE, Long> {
+public interface BoardRepository_save extends JpaRepository<TB_BOARD_SAVE, Integer> {
 
 }
